--- a/Project.toml
+++ b/Project.toml
@@ -1,7 +1,7 @@
 name = "ThinPlateSplines"
 uuid = "1d861738-f48e-4029-b1d3-81ce6bc7f5ab"
 authors = ["Al Nejati"]
-version = "0.1.0"
+version = "0.2.0"
 
 [deps]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"