* Parameterize `ThinPlateSpline`
* Use `hcat` rather than `cat(dims = 2, ...)`