@mkitti 2023-12-07
Type stability in Julia has enormous performance benefits.

# Master branch

```julia
julia> @benchmark tps_deform($x1, $tps)
BenchmarkTools.Trial: 10000 samples with 9 evaluations.
 Range (min … max):  2.944 μs … 472.878 μs  ┊ GC (min … max): 0.00% … 98.75%
 Time  (median):     3.167 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.506 μs ±   6.657 μs  ┊ GC (mean ± σ):  2.65% ±  1.39%

  ██▅▂▆▆▅▄▄▄▃▄▃▃▃▃▃▂▂▁▁▃▂▁    ▃▂▁▁▃▂▁▁▂▂ ▁▂ ▁ ▁               ▂
  ████████████████████████▇▇███████████████████▇██▇▇▇▇▆▇▆▆▅▆▆ █
  2.94 μs      Histogram: log(frequency) by time      5.38 μs <

 Memory estimate: 1.27 KiB, allocs estimate: 37.
```

# This pull request

```julia
julia> @benchmark tps_deform($x1, $tps)
BenchmarkTools.Trial: 10000 samples with 198 evaluations.
 Range (min … max):  439.394 ns …  35.221 μs  ┊ GC (min … max): 0.00% … 94.01%
 Time  (median):     454.040 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   525.449 ns ± 678.367 ns  ┊ GC (mean ± σ):  5.28% ±  4.25%

  █▆▆▃▂▁▂▃▃▃▂ ▄▃▁       ▁                                       ▂
  █████████████████▇▇█▇▇██▇▅▅▄▄▅▄▄▃▁▁▁▃▄▄▃▄▅▃▃▁▄▄▅▃▁▁▃▃▁▁▁▁▁▃▁▆ █
  439 ns        Histogram: log(frequency) by time       1.26 μs <

 Memory estimate: 480 bytes, allocs estimate: 5.
```